+++ frames/research.lua
@@ -1,3 +1,4 @@
+local addonName, addon = ...
 local _G = _G
 local select, unpack, ipairs, pairs, type, tostring, table, print = _G.select, _G.unpack, _G.ipairs, _G.pairs, _G.type,
     _G.tostring, _G.table, _G.print